Refactoring TaskForm to sep. component file
