+++ src/App.js
@@ -1,31 +1,6 @@
 import React, { useState } from "react";
 import Task from "./components/Task";
-
-// component for adding tasks to list
-function TaskForm({ addTask }) {
-  // hooks w/ state for new tasks
-  const [value, setValue] = useState("");
-  // submit handler for new tasks
-  const handleSubmit = e => {
-    e.preventDefault(); // prevent form sub & page reload
-    if (!value) return;
-    addTask(value);
-    setValue("");
-  };
-
-  return (
-    <form onSubmit={handleSubmit}>
-      <input
-        type="text"
-        className="taskForm"
-        value={value} // state value comes in here
-        onChange={e => setValue(e.target.value)} // use the hook to set the new val
-        placeholder="Give me a task. . ."
-        autoFocus
-      />
-    </form>
-  );
-}
+import TaskForm from './components/TaskForm';
 
 // HOC component
 function App() {
